[fix] Fix faulty npm lifecycle configuration
When installing githooks, I misunderstood how this was structured.  I thought I needed the postinstall and preuninstall hooks to wire up githooks.  BUT, that wiring is done by the postinstall and preuninstall hooks **of the githooks** library, making this both redundant and harmful to consumers who don't have githooks installed as part of their project.